aoc-2023/aoc-2023-day-18/solution-in-python3/test_solution.py
import pytest

import solution


@pytest.mark.parametrize(
    "filename,expected",
    [
        pytest.param("puzzle-input-example.txt", 62),
        pytest.param("puzzle-input-full.txt", -1 ), # TODO # 46131 too low ( perimeter_count=3730 content_count=42401)! # 43258 too low
    ],    
)
def test_solve_part1(filename, expected):
    assert solution.solve_part1(filename) == expected

@pytest.mark.parametrize(
    "filename,expected",
    [
        #pytest.param("puzzle-input-example.txt", -1),
        #pytest.param("puzzle-input-full.txt", -1),
    ],    
)
def test_solve_part2(filename, expected):
    assert solution.solve_part2(filename) == expected    